requirements.txt
streamlit
Pillow
PyMuPDF
plotly
python-dotenv
google-generativeai
